add role controller: validation, queryparameters, pagedresponse

+++ Cinema.Server/Cinema.API/Controllers/RoleController.cs
@@ -1,9 +1,9 @@
 ﻿using AutoMapper;
+using Cinema.Application.Common.Models;
 using Cinema.Application.DTOs.RoleDtos;
 using Cinema.Application.Interfaces;
 using Cinema.Domain.Entities;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace Cinema.API.Controllers
 {
@@ -13,20 +13,20 @@ namespace Cinema.API.Controllers
     {
         private readonly IRoleRepository _roleRepository;
 
-        public RolesController(IRoleRepository roleRepository, IMapper mapper):base(mapper)
+        public RolesController(IRoleRepository roleRepository, IMapper mapper) : base(mapper)
         {
             _roleRepository = roleRepository;
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<RoleDto>>> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryParameters queryParameters)
         {
-            var roles = await _roleRepository.GetAll().ToListAsync();
-            return Ok(_mapper.Map<IEnumerable<RoleDto>>(roles));
+            var results = await _roleRepository.GetAllPagedAsync(queryParameters);
+            return OkPaged<Role, RoleDto>(results, queryParameters);
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<RoleDto>> GetById(int id)
+        public async Task<IActionResult> GetById(int id)
         {
             var role = await _roleRepository.GetByIdAsync(id);
             if (role == null) return NotFound();
@@ -35,18 +35,18 @@ namespace Cinema.API.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult<RoleDto>> Create(RoleCreateDto dto)
+        public async Task<IActionResult> Create([FromBody] RoleCreateDto dto)
         {
             var role = _mapper.Map<Role>(dto);
             await _roleRepository.AddAsync(role);
             await _roleRepository.SaveAsync();
 
-            var result = _mapper.Map<RoleDto>(role);
-            return CreatedAtAction(nameof(GetById), new { id = role.RoleId }, result);
+            var response = _mapper.Map<RoleDto>(role);
+            return CreatedAtAction(nameof(GetById), new { id = role.RoleId }, response);
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Update(int id, RoleCreateDto dto)
+        public async Task<IActionResult> Update(int id, [FromBody] RoleCreateDto dto)
         {
             var role = await _roleRepository.GetByIdAsync(id);
             if (role == null) return NotFound();
@@ -58,7 +58,7 @@ namespace Cinema.API.Controllers
         }
 
         [HttpPatch("{id}")]
-        public async Task<IActionResult> Patch(int id, RolePatchDto dto)
+        public async Task<IActionResult> Patch(int id, [FromBody] RolePatchDto dto)
         {
             var role = await _roleRepository.GetByIdAsync(id);
             if (role == null) return NotFound();
@@ -72,8 +72,12 @@ namespace Cinema.API.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var role = await _roleRepository.GetByIdAsync(id);
+            if (role == null) return NotFound();
+
             await _roleRepository.DeleteAsync(id);
             await _roleRepository.SaveAsync();
+
             return NoContent();
         }
     }
